% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junrui.edit.R
\name{ActCosinor}
\alias{ActCosinor}
\title{Cosinor Model for Circadian Rhythmicity for Whole Dataset}
\usage{
ActCosinor(data, window = 1)
}
\arguments{
\item{data}{\code{data.frame} of dimension n*1442 containing the 1440 minute activity data for all n subject days.
The first two columns have to be ID and Day.}

\item{window}{\code{number} 1440/window is the dimension of the data. Default is 1.}
}
\value{
A \code{data.frame} with following columns
\item{ID}{identifier of the person}
\item{mesor}{mesor}
\item{amp}{amplitude} 
\item{acro}{acrophase}
}
\description{
A parametric approach to study circadian rhythmicity assuming cosinor shape.
Here we used the anti logistic transformed cosine curves, which provided 3 parameters.
}
