% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_data2.R
\name{bin_data2}
\alias{bin_data2}
\title{Bin data into longer windows}
\usage{
bin_data2(x = x, window = 1, method = c("average", "sum"))
}
\arguments{
\item{x}{\code{vector} of activity data.}

\item{window}{window size used to bin the original 1440 dimensional data into. Window size
should be an integer factor of 1440}

\item{method}{\code{character} of "sum" or "average", function used to bin the data}
}
\value{
a vector of binned data
}
\description{
Bin minute level data into different time resolutions
}
