% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS2.R
\name{IS2}
\alias{IS2}
\title{Interdaily Statbility}
\usage{
IS2(x)
}
\arguments{
\item{x}{\code{data.frame} of dimension ndays by p, where p is the dimension of the data.}
}
\description{
This function calcualte interdaily stability, a nonparametric metric
of circadian rhtymicity
}
\references{
Junrui Di et al. Joint and individual representation of domains of physical activity, sleep, and circadian rhythmicity. Statistics in Biosciences.
}
