% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junrui.edit.R
\name{IV_long2}
\alias{IV_long2}
\title{Intradaily Variability}
\usage{
IV_long2(count.data, level = 600)
}
\arguments{
\item{count.data}{\code{data.frame} of dimension n * 1442 containing the 1440 minute activity data for all n subject days.}

\item{level}{\code{character} the number of seconds in each window. level= 600 seconds (10 minutes) by default.}
}
\value{
IV
}
\description{
This function calcualte intradaily variability, a nonparametric metric
reprsenting fragmentation of circadian rhtymicity
}
