% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.jive.predict.edit.R
\name{jive.predict2}
\alias{jive.predict2}
\title{Modified jive.predict function  (package: r.jive)}
\usage{
jive.predict2(data.new, jive.output)
}
\arguments{
\item{data.new}{\code{data.new}  A list of two or more linked data matrices on which to estimate JIVE scores.  These matrices must have the same column dimension N, which is assumed to be common.}

\item{jive.output}{\code{jive.output} An object of class "jive", with row dimensions matching those for data.new.}
}
\value{
See r.jive:: jive.predict for details
}
\description{
Replace SVDmiss by SVDmiss2 in the function
}
\details{
See jive.predict(package:r.jive) for details.
}
