% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA2.R
\name{RA2}
\alias{RA2}
\title{Relative Amplitude}
\usage{
RA2(x, window = 1, method = c("average", "sum"))
}
\arguments{
\item{x}{\code{vector} vector of activity data}

\item{window}{since the caculation of M10 and L5 depends on the dimension of data, we need to include
window size as an argument.}

\item{method}{\code{character} of "sum" or "average", function used to bin the data}
}
\value{
RA
}
\description{
This function calcualte relative amplitude, a nonparametric metric
reprsenting fragmentation of circadian rhtymicity
}
\references{
Junrui Di et al. Joint and individual representation of domains of physical activity, sleep, and circadian rhythmicity. Statistics in Biosciences.
}
