% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time2.R
\name{Time_long2}
\alias{Time_long2}
\title{Timne Metrics for Whole Dataset}
\usage{
Time_long2(count.data, weartime, thresh, smallerthan = TRUE, bout.length = 1)
}
\arguments{
\item{count.data}{\code{data.frame} of dimension n*1442 containing the 1440 minute activity data for all n subject days.
The first two columns have to be ID and Day.}

\item{weartime}{\code{data.frame} with dimension of \code{count.data}.
The first two columns have to be ID and Day.}

\item{thresh}{threshold to binarize the data.}

\item{smallerthan}{Find a state that is smaller than a threshold, or greater than or equal to.}

\item{bout.length}{minimum duration of defining an active bout; defaults to 1.}
}
\value{
A dataframe with some of the following columns
\item{ID}{identifier of the person}
\item{Day}{indicator of which day of activity it is, can be a numeric vector of sequence 1,2,... or a string of date}
\item{time}{time of certain state}
}
\description{
This function is a whole dataset wrapper for \code{Time}
}
