% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS.R
\name{IS_long2}
\alias{IS_long2}
\title{Interdaily Statbility}
\usage{
IS_long2(count.data, level = "hour")
}
\arguments{
\item{count.data}{\code{data.frame} of dimension p by 1440, representing the p days of activity data for one subjectect}

\item{level}{\code{character} level="hour" or not}
}
\value{
IS
}
\description{
This function calcualte interdaily stability, a nonparametric metric of circadian rhtymicity
}
