% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junrui.edit.R
\name{RA2}
\alias{RA2}
\title{Relative Amplitude}
\usage{
RA2(x)
}
\arguments{
\item{x}{\code{vector} vector of dimension 440 which reprsents 1440 minute activity data}
}
\value{
RA
}
\description{
This function calcualte relative amplitude, a nonparametric metric
reprsenting fragmentation of circadian rhtymicity
}
